+++ apps/web/src/app/organizations/manage/collections.component.ts
@@ -20,9 +20,8 @@ import {
 import { ListResponse } from "@bitwarden/common/models/response/listResponse";
 import { CollectionView } from "@bitwarden/common/models/view/collectionView";
 
-import { EntityUsersComponent } from "../../modules/organizations/manage/entity-users.component";
-
 import { CollectionAddEditComponent } from "./collection-add-edit.component";
+import { EntityUsersComponent } from "./entity-users.component";
 
 @Component({
   selector: "app-org-manage-collections",
