+++ apps/web/src/app/shared/loose-components.module.ts
@@ -28,9 +28,9 @@ import { FooterComponent } from "../layouts/footer.component";
 import { FrontendLayoutComponent } from "../layouts/frontend-layout.component";
 import { NavbarComponent } from "../layouts/navbar.component";
 import { UserLayoutComponent } from "../layouts/user-layout.component";
-import { OrganizationCreateModule } from "../modules/organizations/create/organization-create.module";
 import { VaultFilterModule } from "../modules/vault-filter/vault-filter.module";
 import { OrganizationBadgeModule } from "../modules/vault/modules/organization-badge/organization-badge.module";
+import { OrganizationCreateModule } from "../organizations/create/organization-create.module";
 import { OrganizationLayoutComponent } from "../organizations/layouts/organization-layout.component";
 import { BulkConfirmComponent as OrgBulkConfirmComponent } from "../organizations/manage/bulk/bulk-confirm.component";
 import { BulkRemoveComponent as OrgBulkRemoveComponent } from "../organizations/manage/bulk/bulk-remove.component";
